Improving the handling of -Xthing:value options to prevent duplicates
Separating the option on a possible : and setting the key to Xthing:
and value to the remaining allows for the args.rb to maintain the
hash of these options.

This has the added benefit of always setting @value to be something
even if that is an empty string, thus the to_s method can be simplified
greatly.